requirements.txt
# Core ML Libraries
tensorflow>=2.15.0
keras>=2.15.0
numpy>=1.24.3
pandas>=2.1.0
scikit-learn>=1.3.0

# Data Acquisition & Analysis
yfinance>=0.2.36
matplotlib>=3.8.0
seaborn>=0.13.0
scipy>=1.11.0

# Documentation & Utilities
jupyter>=1.0.0
pypdf>=4.0.0
pdfminer.six>=20231228
pymupdf>=1.23.0
pdfplumber>=0.10.0
